curie-scientific-research-ai-assistant/requirements.txt
streamlit
transformers
sentence-transformers
PyMuPDF
faiss-cpu
scikit-learn
dogs